[#step3] refactor: 뷰가 머니 생성 객체를 반환하고 머니 생성 객체가 머니를 반환하는 구조로 변경
	이유: 뷰와 모델을 분리하기 위해 모델을 생성하는 객체를 만들었고 이를 적용하였다.
