+++ VendingMachineApp/VendingMachineApp/Models/Money/Instantiators/MoneyInstantiator.swift
@@ -1,0 +1,13 @@
+//
+//  MoneyInstantiator.swift
+//  VendingMachineApp
+//
+//  Created by kimdo2297 on 2020/06/24.
+//  Copyright © 2020 Jason. All rights reserved.
+//
+
+import Foundation
+
+class MoneyInstantiator {
+    var money: Money? { return nil }
+}
